refactor: break down big component in smaller ones
