+++ src/features/ExpendedNote/ui/Content.tsx
@@ -1,0 +1,23 @@
+import { useState } from 'react';
+
+import selectActiveNote from '@pages/Notes/lib/selectors/selectActiveNote';
+import { mockedNotes } from '@shared/lib/const';
+import useAppSelector from '@shared/lib/hooks/useAppSelector';
+
+const Content = () => {
+  const selectedNote = useAppSelector(selectActiveNote);
+  const { content: initialContent } = mockedNotes[selectedNote];
+
+  const [content, setContent] = useState(initialContent);
+
+  return (
+    <textarea
+      placeholder="Start writing your notes here :)"
+      className="h-full w-full resize-none whitespace-pre rounded-sm bg-transparent px-3 py-1 text-on-surface-variant outline-none transition-all focus:ring-4 focus:ring-high-contrast-inverse-primary focus:ring-offset-4 focus:ring-offset-surface"
+      value={content}
+      onChange={(e) => setContent(e.target.value)}
+    />
+  );
+};
+
+export default Content;
